coulomb rupture set follows runner->configuration->execute pattern
